refactor: decompose config.ts and review-reducer.ts
Both files were absorbing every new concern in their area and had grown
past the point where they can be scanned.

config.ts (1117 lines) split at its natural seam: the declarative
`.kodiai.yml` schema -- the part contributors actually read when adding a
setting -- was buried above several hundred lines of parsing,
sanitization, and fail-open warning logic. Schemas move to
config-schema.ts, leaving the loader at 480 lines. Only three symbols
crossed the boundary.

review-reducer.ts (858 lines) sheds two cohesive clusters: the shared
pipeline types, so the reducer, its summary projection, and its callers
can agree on a contract without importing the implementation; and the
details-summary projection, which has a requirement the reducer does not
share -- everything it emits is PR-visible, so tokens are sanitized and
the string is length-bounded. That redaction boundary is now obvious
instead of trailing a 900-line module. Re-exports keep one entry point,
so no caller changed.

Pure moves: no test file needed updating except
deploy-timeout-alignment.test.ts, which greps config source by path and
correctly caught the relocation.

Also drops a `js-yaml` import left dead in the extracted schema module.
Worth noting separately: tsconfig sets `noUnusedLocals: false`, which is
why that import and an earlier dead helper both passed typecheck and
lint. Turning it on surfaces ~175 pre-existing violations, so it is left
alone here rather than ballooning this change.
